Initial commit.
abundance_table_transpose is a tool to take a list of abundance and
functional information by ID and turn it into a table with function
as rows and IDs as columns. By default all the functional data is kept,
but users can choose to keep less information (levels 1-3 and function),
in which case ID abundance will be aggregated.