Makes several changes to mgrast_request().
Allows for optional item_id and API parameters, and parameters with lists
as values. Previously the item_id parameter was required, but not all API
calls use an ID. This change makes it optional, and only adds a preceding
forward-slash to the request URL if the item_id is present.

The same problem existed with the API parameter list. Now, if no parameters
are present, there is no '?' in the URL.

The final change allows for values in the parameter dict to be lists instead
of single values. The main use case for this is with the 'matrix' API call.
This call allows for multiple 'id' entries in the URL parameter list, e.g.
http://.../?id=a&id=b&id=c. This change detects lists in a parameter value
and creates an appropriate string representation for the request URL.