download_stage now skips previously downloaded files.
After file information is retrieved, the program checks whether the file already exists and if the file is the same size as reported by the MG-RAST server. If both checks pass, the file is skipped and the user is notified.

This commit also adds a new program option '--force' that overrides the above checks and downloads the file anyway.